requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# Async HTTP Client
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0

# XML Processing
lxml==4.9.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
mypy==1.7.1
